inputValidator refactored
