+++ src/routes/auth.ts
@@ -12,9 +12,9 @@ import BadRequestError from '../exceptions/BadRequestError';
 import Admin from '../datatypes/authentication/Admin';
 import AdminSession from '../datatypes/authentication/AdminSession';
 import LoginCredentials from '../datatypes/authentication/LoginCredentials';
-import {validateLoginCredentials} from '../functions/inputValidator/validateLoginCredentials';
-import checkUsernameRule from '../functions/inputValidator/checkUsernameRule';
-import checkPasswordRule from '../functions/inputValidator/checkPasswordRule';
+import {validateLoginCredentials} from '../functions/inputValidator/admin/validateLoginCredentials';
+import checkUsernameRule from '../functions/inputValidator/admin/checkUsernameRule';
+import checkPasswordRule from '../functions/inputValidator/admin/checkPasswordRule';
 import createAccessToken from '../functions/JWT/createAccessToken';
 import createRefreshToken from '../functions/JWT/createRefreshToken';
 import verifyRefreshToken from '../functions/JWT/verifyRefreshToken';
